Add user authentication UI
popup.js now renders the firebaseui for login options. Signup has not
been implemented yet. Once the user logs in, the firebase object in
background.js and options.js both are updated with the credentials of
the logged in user. The information pertaining to the current user is
stored in firebase.auth().currentUser
Also, I've updated my firebase with auth rules so accessing user_0 no
longer works if you're using the credentials for my firebase. We'll work
on generalizing the functions that rely on user_0 next.