Fix domain updates and revamp UI rendering
I made a bunch of changes in the last three commits (including this
    one). First, there's the project directory restructuring. It's
fairly standard stuff, although know that the order of js scripts that
are to be executed is defined in the respective html file. Furthermore,
I've made sure that all the firebase communications goes through
background.js (api.js houses those functions now). See
background_comm.js to get a sense of how to use the messaging API to
facilitate this communication. Lastly, I've setup a simple system of
HTML rendering using refresh() (in popup.js). The idea is that a global
state is stored at the top of popup.js (userContext, currentScreen,
    etc.). Whenever these variables are updated, we mandate a call to
refresh(). This will then check those globals and update the popup UI
accordingly.